+++ SpiceQL/tests/Fixtures.cpp
@@ -175,7 +175,7 @@ void IsisDataDirectory::compareKernelSets(string name, set<string> expectedDiff)
 }
 
 
-void IsisDataDirectory::CompareKernelSets(vector<string> kVector, vector<string> expectedSubSet) {
+void IsisDataDirectory::compareKernelVector(vector<string> kVector, vector<string> expectedSubSet) {
   for (auto &e : kVector) { 
     auto it = find(kVector.begin(), kVector.end(), e);
     if (it == kVector.end()) {